Check if user has any regex
git-svn-id: svn://svn.newznab.com/nn/trunk@1003 a8b8df2f-1e19-4095-b43c-2e6d618ca476

+++ www/admin/regex-submit.php
@@ -10,59 +10,67 @@ $page->title    = "Submit your regex expressions to the Official Database";
 // Logic
 $regex          = new ReleaseRegex();
 $regexList      = $regex->get(false, -1, true, true);
-$regexSerialize = serialize($regexList);
-$regexFilename  = 'releaseregex-' . time() . '.regex';
 
-// User wants to submit their regex's
-if (isset($_POST['regex_submit_please']))
+if (count($regexList))
 {
-  // Delete old regex files
-  $oldFiles = glob(WWW_DIR . '/temp/*.regex') ? glob(WWW_DIR . '/temp/*.regex') : array();
-  foreach ($oldFiles as $oldRegexFilename) @unlink($oldRegexFilename);
+  $regexSerialize = serialize($regexList);
+  $regexFilename  = 'releaseregex-' . time() . '.regex';
 
-  // Create new regex file
-  if (file_put_contents(WWW_DIR . "/temp/$regexFilename", $regexSerialize))
+  // User wants to submit their regex's
+  if (isset($_POST['regex_submit_please']))
   {
-    // Continue processing
-    if (file_exists(WWW_DIR . "/temp/$regexFilename") &&
-        is_readable(WWW_DIR . "/temp/$regexFilename"))
+    // Delete old regex files
+    $oldFiles = glob(WWW_DIR . '/temp/*.regex') ? glob(WWW_DIR . '/temp/*.regex') : array();
+    foreach ($oldFiles as $oldRegexFilename) @unlink($oldRegexFilename);
+
+    // Create new regex file
+    if (file_put_contents(WWW_DIR . "/temp/$regexFilename", $regexSerialize))
     {
-      // Submit
-
-      $ch = curl_init();
-      curl_setopt($ch, CURLOPT_HEADER, 0);
-      curl_setopt($ch, CURLOPT_VERBOSE, 0);
-      curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-      curl_setopt($ch, CURLOPT_USERAGENT, "Mozilla/4.0 (compatible;)");
-      curl_setopt($ch, CURLOPT_URL,"http://newznab.com/regex/upload.php");
-      curl_setopt($ch, CURLOPT_POST, true);
-      $post = array(
-          "regex_file" => "@" . WWW_DIR . "/temp/$regexFilename",
-      );
-      curl_setopt($ch, CURLOPT_POSTFIELDS, $post); 
-      $response = curl_exec($ch);
-
-      echo curl_error($ch);
-
-      curl_close($ch);
-
-      if ($response == 'OK')
+      // Continue processing
+      if (file_exists(WWW_DIR . "/temp/$regexFilename") &&
+          is_readable(WWW_DIR . "/temp/$regexFilename"))
       {
-        $page->smarty->assign('upload_status', 'OK');
+        // Submit
 
-        // Remove old regex file
-        @unlink(WWW_DIR . "/temp/$regexFilename");
+        $ch = curl_init();
+        curl_setopt($ch, CURLOPT_HEADER, 0);
+        curl_setopt($ch, CURLOPT_VERBOSE, 0);
+        curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+        curl_setopt($ch, CURLOPT_USERAGENT, "Mozilla/4.0 (compatible;)");
+        curl_setopt($ch, CURLOPT_URL,"http://newznab.com/regex/upload.php");
+        curl_setopt($ch, CURLOPT_POST, true);
+        $post = array(
+            "regex_file" => "@" . WWW_DIR . "/temp/$regexFilename",
+        );
+        curl_setopt($ch, CURLOPT_POSTFIELDS, $post); 
+        $response = curl_exec($ch);
+
+        curl_close($ch);
+
+        if ($response == 'OK')
+        {
+          $page->smarty->assign('upload_status', 'OK');
+
+          // Remove old regex file
+          @unlink(WWW_DIR . "/temp/$regexFilename");
+        }
+        else
+        {
+          $page->smarty->assign('upload_status', 'BAD');
+        }
       }
       else
       {
-        $page->smarty->assign('upload_status', 'BAD');
+        $regexFilename = 'Unable to generate file! Please try again.';
       }
     }
-    else
-    {
-      $regexFilename = 'Unable to generate file! Please try again.';
-    }
   }
+}
+else
+{
+  $regexFilename  = 'No user regexs found. Please add some.';
+  $regexList      = array('Empty');
+  $page->smarty->assign('regex_error', 1);
 }
 
 // Assigns
